Library/Mobile Documents/com~apple~CloudDocs/selenium_python_automation_practice_oct25/requirements.txt
attrs==25.4.0
certifi==2025.10.5
charset-normalizer==3.4.3
h11==0.16.0
idna==3.10
outcome==1.3.0.post0
packaging==25.0
PySocks==1.7.1
python-dotenv==1.1.1
requests==2.32.5
selenium==4.36.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.31.0
trio-websocket==0.12.2
typing_extensions==4.15.0
urllib3==2.5.0
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
pytest==8.4.2
selenium-page-factory==2.7.0
allure-pytest==2.15.0
